requirements.txt
PyMySQL==0.9.2
asn1crypto==0.24.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
idna==2.7
lxml==4.2.5
pycparser==2.19
requests==2.19.1
six==1.11.0
urllib3==1.23
setuptools
mysql-connector-python
